+++ CHANGES.rst
@@ -13,6 +13,7 @@ Additions
 - consistent_read parameter to ``Query``
 - central config.py file with all constraints
 - timer for table status changes
+- ``Query`` supports ``Limit`` and ``ScanIndexForward`` parameters
 
 Removal
 -------